long string refactoring

+++ src/main/java/birsy/clinker/common/page/elements/ImagePageElement.java
@@ -15,15 +15,15 @@ public class ImagePageElement extends PageElement {
     public static final MapCodec<ImagePageElement> CODEC = RecordCodecBuilder.mapCodec(
             instance -> instance.group(
                     ResourceLocation.CODEC.fieldOf("texture").forGetter(element -> element.texture),
-                    Codec.DOUBLE.listOf(4, 4).optionalFieldOf("uv_coordinates", List.of(0.0, 0.0, 1.0, 1.0)).forGetter(element -> List.of(element.u1, element.v1, element.u2, element.v2)),
+                    Codec.FLOAT.listOf(4, 4).optionalFieldOf("uv_coordinates", List.of(0.0F, 0.0F, 1.0F, 1.0F)).forGetter(element -> List.of(element.u1, element.v1, element.u2, element.v2)),
                     PageElementTransform.CODEC.fieldOf("transform").forGetter(element -> element.transform)
             ).apply(instance, ImagePageElement::new)
     );
 
     final ResourceLocation texture;
-    final double u1, v1, u2, v2;
+    final float u1, v1, u2, v2;
 
-    public ImagePageElement(ResourceLocation texture, List<Double> textureCoordinates, PageElementTransform transform) {
+    public ImagePageElement(ResourceLocation texture, List<Float> textureCoordinates, PageElementTransform transform) {
         super(transform);
         this.texture = texture;
         this.u1 = textureCoordinates.get(0); this.v1 = textureCoordinates.get(1);
